Revert "Update screenshots"
This reverts commit 0d006c223c8093f5d1d215180d98ff80fd738ec2.
